my-notebook/requirements.txt
langchain==0.3.0
langchain-community==0.0.27
langchain-openai==0.0.8
langchain-chroma==0.0.1
openai==1.14.0
chromadb==0.4.24
python-dotenv==1.0.1
pypdf==4.1.0
requests==2.31.0
beautifulsoup4==4.12.3
sqlite3==3.45.1
rich==13.7.0 